added containerSettingId setting
It seems very difficult to add container settings to the editor.
as of right now it seems that doing so requires a mixture of
setting containerSettingEnabled to true, making sure that
containerSettingInitFunction is a fuction and then using the
containerSettingShowFunction and containerSettingHideFunction
to hide and show your setting container.

By passing an Id with this setting, your form will automatically
be injected and handled properly.