add setContent to keditor.js
This is an initial setContent.  It reformats content nicely to the original.    The current and major problem is that I am not able to drop containers and components to the existing content area.  I notice that when I drop a container onto an existing container I am able to drop components to that particular container.  I am not able to drop new containers. Kindly advise